+++ lib/cards/CurrentBalanceCard.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:firebase_database/firebase_database.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 // import 'database.dart';
-import 'SetBalancePage.dart';
+import '../screens/SetBalancePage.dart';
 
 class CurrentBalanceCard extends StatefulWidget {
   const CurrentBalanceCard({Key? key}) : super(key: key);
