+++ lib/screens/SetBalancePage.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'database.dart';
+import '../others/database.dart';
 
 class SetBalancePage extends StatefulWidget {
   @override
